requirements.txt
pyenchant
python-Levenshtein
gensim
scikit-learn
nltk
pandas
openpyxl